// src/file/enum/file.enum.ts
export enum FileStatusEnum {
  Add = 'Add',
  Remove = 'Remove',
}

export enum FileTypeEnum {
  Video = 'Video',
  Audio = 'Audio',
  Document = 'Document',
  Image = 'Image',
  Svg = 'image/svg+xml',
}
export enum ACLEnum {
  PublicRead = 'public-read',
  Private = 'private',
}

export enum FilePathEnum {
  Host = 'host',
  Guest = 'guest',
}